% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sex_age_specific_file.R
\name{add_country_province_name}
\alias{add_country_province_name}
\title{Function to add full names of provinces and country to data}
\usage{
add_country_province_name(data)
}
\arguments{
\item{data}{data with country_province column}
}
\value{
cleaned dataset with country_province_name column
}
\description{
Function to add full names of provinces and country to data
}
\examples{
\dontrun{
  add_country_province_name(data)
}

}
