% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_formatting_label_cells}
\alias{add_formatting_label_cells}
\title{returns "normal" or "percentage" depending on format of current year data}
\usage{
add_formatting_label_cells(a_file, a_sheet = "SA", a_col = 43)
}
\arguments{
\item{a_file}{Excel file created by Thembisa Project}

\item{a_sheet}{sheet name, default is "SA"}

\item{a_col}{column number to check for formatting, default is 43 (column AQ)}
}
\value{
a tibble with indicators and their formatting type
}
\description{
returns "normal" or "percentage" depending on format of current year data
}
\examples{
\dontrun{
  add_formatting_label_cells(a_file, a_sheet = "SA", a_col = 43)
}

}
