% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_formatting_label_keywords}
\alias{add_formatting_label_keywords}
\title{returns "normal" or "percentage" depending on keyword of an indicator}
\usage{
add_formatting_label_keywords(a_file, a_sheet = "SA")
}
\arguments{
\item{a_file}{Excel file created by Thembisa Project}

\item{a_sheet}{sheet name, default is "SA"}
}
\value{
a tibble with indicators and their formatting type
}
\description{
returns "normal" or "percentage" depending on keyword of an indicator
}
\examples{
\dontrun{
  add_formatting_label_keywords(a_file, a_sheet = "SA")
}

}
