% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_prov_output.R
\name{clean_prov_output}
\alias{clean_prov_output}
\title{Clean provincial output}
\usage{
clean_prov_output(nat_prov_sheet)
}
\arguments{
\item{nat_prov_sheet}{a sheet from the provincial output excel file}
}
\value{
cleaned dataset
}
\description{
Clean provincial output
}
\examples{
\dontrun{
  clean_prov_output(nat_prov_sheet)
}

}
