% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sex_age_specific.R
\name{clean_sex_age_specific}
\alias{clean_sex_age_specific}
\title{clean data from the sex_age specific file}
\usage{
clean_sex_age_specific(excel_tab)
}
\arguments{
\item{excel_tab}{a tab in excel}
}
\value{
cleaned dataset
}
\description{
clean data from the sex_age specific file
}
\examples{
\dontrun{
  clean_sex_age_specific(excel_tab)
}

}
