% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_percentage}
\alias{is_percentage}
\title{Returns if a cell is formatted as percentage}
\usage{
is_percentage(wb, sheet, row, col = 43)
}
\arguments{
\item{wb}{workbook object}

\item{sheet}{sheet name}

\item{row}{row number (0-indexed)}

\item{col}{column number (1-indexed, default is 43 which is column AQ)}
}
\value{
TRUE if the cell is formatted as percentage, FALSE otherwise
}
\description{
Returns if a cell is formatted as percentage
}
\examples{
\dontrun{
  is_percentage(wb, sheet, row, col = 43)
}

}
