% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sex_age_specific_file.R
\name{read_file}
\alias{read_file}
\title{Generic reading function for Thembisa excel files}
\usage{
read_file(filename, sheets_to_exclude = "Notes")
}
\arguments{
\item{filename}{path to the Thembisa excel file}

\item{sheets_to_exclude}{sheets to exclude from reading, default is "Notes"}
}
\value{
cleaned dataset
}
\description{
Generic reading function for Thembisa excel files
}
\examples{
\dontrun{
  read_file(filename, sheets_to_exclude)
}

}
