% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_prov_output}
\alias{read_prov_output}
\title{Generic reading function for Thembisa excel files}
\usage{
read_prov_output(
  filename,
  sheets_to_exclude = c("Notes", "Comparison", "Abbreviations")
)
}
\arguments{
\item{filename}{path to the Thembisa excel file}

\item{sheets_to_exclude}{sheets to exclude from reading, default is Notes, Comparison and Abbreviations}
}
\value{
cleaned dataset
}
\description{
Generic reading function for Thembisa excel files
}
\examples{
\dontrun{
  read_prov_output(filename, sheets_to_exclude)
}

}
