% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sex_age_specific_file.R
\name{read_sex_age_specific_file}
\alias{read_sex_age_specific_file}
\title{Read in all data from Thembisa excel file}
\usage{
read_sex_age_specific_file(filename)
}
\arguments{
\item{filename}{path to the Thembisa excel file}
}
\value{
cleaned dataset
}
\description{
Read in all data from Thembisa excel file
}
\examples{
\dontrun{
  read_sex_age_specific_file(filename)
}

}
