% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_sex_age_specific_file}
\alias{read_sex_age_specific_file}
\title{Read in all data from Thembisa excel file}
\usage{
read_sex_age_specific_file(filename, sheets_to_exclude = "Notes")
}
\arguments{
\item{filename}{path to the Thembisa excel file}

\item{sheets_to_exclude}{sheets to exclude from reading, default is "Notes"}
}
\value{
cleaned dataset
}
\description{
Read in all data from Thembisa excel file
}
\examples{
\dontrun{
  read_sex_age_specific_file(filename, sheets_to_exclude)
}

}
